requirements.txt
Flask
jinja2
PyYAML
pdfplumber
pillow
pytesseract
rapidfuzz
pint
jsonschema
requests
werkzeug
gunicorn
